refactor: split types.ts into per-layer files (shim) (#632 segment 2 / A1)
Step A1 of types split: extract type definitions from
src/legacy/core/types.ts into per-layer files. The original file becomes
a pure re-export shim — zero behavior change, no callers updated yet.

Per-layer destinations:

- src/agent/types.ts       — ProviderConfig, AgentConfig, CompactionMode, CompactionConfig
- src/sessions/types.ts    — Session, SessionMetadata, SessionConfig, SessionStoreConfig, SessionStore
- src/gateway/types.ts     — PeerKind, BindingPeer, BindingMatch, Binding, ChannelsConfig, Transport
- src/tools/types.ts       — Tool, AgentToolSettings (creates src/tools/)
- src/automation/types.ts  — CronActionConfig (creates src/automation/)
- src/legacy/core/types.ts — re-export shim only

SDK re-exports (AgentMessage, AgentEvent, Usage, ImageContent) stay in
the shim for now; they'll move to src/agent/types.ts in step A2 along
with the SandboxConfig coupling cleanup.

Acknowledged transitional leak: src/agent/types.ts imports SandboxConfig
from ../legacy/sandbox/config.js. Sandbox migration is a later segment;
the leaked path makes the dependency visible.

Step A2 (next PR) will rewrite all importers to point at the new
locations and delete the shim.

tsc clean; 1165/1165 tests pass.
